Refactor tablet_sink file into multi files
