+++ be/src/exec/CMakeLists.txt
@@ -31,6 +31,9 @@ set(EXEC_FILES
     olap_common.cpp
     tablet_info.cpp
     tablet_sink.cpp
+    tablet_sink_colocate_sender.cpp
+    tablet_sink_index_channel.cpp
+    tablet_sink_sender.cpp
     mysql_scanner.cpp
     es/es_predicate.cpp
     es/es_scan_reader.cpp
